+++ glotaran/dataio/wavelength_time_explicit_file.py
@@ -1,5 +1,5 @@
 from glotaran.model import Dataset
-from glotaran.models.spectral_temporal.dataset import SpectralTemporalDataset
+from glotaran.models.spectral_temporal.spectral_temporal_dataset import SpectralTemporalDataset
 from .spectral_timetrace import SpectralUnit
 # this import was unused and flake8 did complain, will leave it as comment
 # from .spectral_timetrace import SpectralTimetrace
