+++ glotaran/models/spectral_temporal/kinetic_megacomplex.py
