+++ glotaran/models/spectral_temporal/kinetic_result.py
