+++ glotaran/models/spectral_temporal/spectral_temporal_dataset_descriptor.py
